src: New message format.
This approach tells the user what's happening rather than leaving the console temporarily blank. Some of these changes still need testing.

+++ src/KryptorCLI/FileEncryption/DecryptFile.cs
@@ -52,10 +52,7 @@ namespace KryptorCLI
             catch (Exception ex) when (ExceptionFilters.Cryptography(ex))
             {
                 CryptographicOperations.ZeroMemory(dataEncryptionKey);
-                if (!(ex is ArgumentException))
-                {
-                    FileHandling.DeleteFile(outputFilePath);
-                }
+                if (ex is not ArgumentException) { FileHandling.DeleteFile(outputFilePath); }
                 throw;
             }
         }
@@ -84,6 +81,7 @@ namespace KryptorCLI
 
         private static void Finalize(string inputFilePath, string outputFilePath, int fileNameLength)
         {
+            Globals.SuccessfulCount += 1;
             RestoreFileName.RenameFile(outputFilePath, fileNameLength);
             FileHandling.DeleteFile(inputFilePath);
         }
